sfn-comprehend-sdk/requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.aws-stepfunctions>=1.132.0
aws-cdk.aws-stepfunctions-tasks>=1.132.0
aws-cdk.assertions>=1.132.0
